add clippy to .husky/pre-commit
- add cargo clippy --all-targets -- -D warnings to .husky/pre-commit
- clear every warning it found

A check that does not fail is not a check, so the hook denies every warning
rather than printing them: ones nobody has to answer for accumulate until the
one that matters is somewhere in the middle of them. The tree was at 48, none
of which had been looked at.

Most were mechanical and the diff carries them. Three are the lint being
wrong about this code, and are answered where the code is rather than by
turning anything off wholesale:

- a test whose assertions compare two constants, which is the point of it: what
  it pins is a relationship between them, and no run can tell you that has been
  edited apart.
- `Judgement::Out`, which nothing in play constructs because the key that did
  was taken away. `Chapters::judge` still knows how to do it and the tests hold
  it to that, and taking a boundary out of the table is something a judging pass
  has to be able to do — the missing half is a key to bind, not the variant.
- mingw resolving `DllMain` to `_DllMain@12`. That fixup is what makes the DLL's
  entry point work and no spelling of the export avoids it.

`hook::address` is new for the twenty-one places that hand a function's address
to a hook as a number. A function item is a type of its own with no address
until it is coerced, so `f as usize` asks for the address of something that has
not got one yet and gets it by an implied coercion — which is what the lint is
about. One name with the reason beside it, rather than the same double cast
written out twenty-one times.

+++ crates/orb/src/text.rs
@@ -56,11 +56,11 @@ impl Font {
             unsafe { std::mem::zeroed() };
         description.lfHeight = -height;
         description.lfWeight = FW_NORMAL;
-        description.lfCharSet = SHIFTJIS_CHARSET as u8;
-        description.lfOutPrecision = OUT_TT_PRECIS as u8;
-        description.lfClipPrecision = CLIP_DEFAULT_PRECIS as u8;
-        description.lfQuality = ANTIALIASED_QUALITY as u8;
-        description.lfPitchAndFamily = DEFAULT_PITCH as u8;
+        description.lfCharSet = SHIFTJIS_CHARSET;
+        description.lfOutPrecision = OUT_TT_PRECIS;
+        description.lfClipPrecision = CLIP_DEFAULT_PRECIS;
+        description.lfQuality = ANTIALIASED_QUALITY;
+        description.lfPitchAndFamily = DEFAULT_PITCH;
         for (slot, unit) in description
             .lfFaceName
             .iter_mut()
